Release 3.6.1

+++ src/app/cli/db.js
@@ -1,12 +1,17 @@
 import { promises as fs } from 'fs'
 import { join } from 'path'
+import { fileExists } from '../../helpers'
 import useIO from "../composables/useIO";
 
 export default function (vorpal, { config, restart }) {
   const dropDatabase = async () => {
     const { deleteAllUploadedFiles } = useIO()
     const dbFile = join(config.root, config.database)
-    await fs.rm(dbFile)
+
+    if (await fileExists(dbFile)) {
+      await fs.rm(dbFile)
+    }
+    
     await deleteAllUploadedFiles()
     return restart()
   }
